#83 add invoice payment form

+++ BionicInventory.Domain/Invoice/InvoicePayment/InvoicePayment.cs
@@ -3,7 +3,7 @@
  * @Author:  Mikael Araya
  * @Contact: MikaelAraya12@gmail.com
  * @Last Modified By:  Mikael Araya
- * @Last Modified Time: Nov 6, 2018 8:46 PM
+ * @Last Modified Time: Nov 12, 2018 12:57 AM
  * @Description: Modify Here, Please 
  */
 using System;
@@ -14,7 +14,7 @@ namespace BionicInventory.Domain.Invoices.InvoicePayment {
     public class InvoicePayments {
         public uint Id { get; set; }
         public uint InvoiceNo { get; set; }
-        public float Amount { get; set; }
+        public double Amount { get; set; }
         public DateTime? DateAdded { get; set; }
         public DateTime? DateUpdated { get; set; }
         public uint CashierId { get; set; }
